MutationObserver now used to re-add button
With MutationObserver, the issue where the button is sometimes not displayed should now be fixed. As such, other code that was used to call addButton (or init) is now redundant and has been removed, with the exception of the "play" event listener (just in case).

removeButton() has also been removed as it serves no purpose.

...also, it turns out the issue that was supposed to be fixed in the previous update still occurred, so I tried fixing the fixed fix by 1) making sure 'vidNum' is up to date and 2) reordering terms in the boolean expression so that it should be evaluated faster if it is false.
(for reference: when the button is enabled while on the first video in a playlist, the script may cause the browser to freeze at the end of the video. I assume this has something to do with repeated calls to checkTime (or something inside it).)