Split up @match in two
Once again, `@match` seems to be treated differently by different userscript managers; in this case, the issue appears to be that  `*`  does not seem to work as a wildcard in the middle of a term for some. As such, `@match http*:` has been split up into `@match http:` and `@match https:` (assuming that all userscript managers allow multiple match tags).
Thanks to Greasy Fork user johnyept for reporting this!