Add try/catch to checkIncomplete, replace innerHTML
A Greasy Fork user reported an issue where `checkIncomplete()` might cause a TypeError "Cannot read properties of undefined (reading 'innerHTML')". Note however that I was unable to replicate this, so I'm not sure whether this is an error that happens occasionally (e.g. because the playlist takes too long to load) or consistently (e.g. due to a UI change introduced through Youtube's A/B testing). Either way, worst-case scenario the script should continue working at the expense of potentially inaccurate times/percentages not being indicated as such (through `>` or `~` symbols by default).

I also took the opportunity to replace `innerHTML` by the safer `innerText`, and added a debug function (which currently is only called in checkIncomplete if the aforementioned issue occurs again).