Fix selectors, add debug messages
So it seems the issue mentioned in the previous commit (57c4b6b363839abf6b219142e3cb5fd78a226f82) was indeed caused by Youtube's A/B testing, since as of yesterday opening Youtube in an incognito window caused the script to no longer work (for me). In addition, the attribute that indicates whether the miniplayer is active was changed (again).

More messages are now also printed if `debug` is enabled.